+++ CHANGELOG.md
@@ -4,15 +4,24 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-01-04
+
 ### Added
+
 - Automatically fetching data
 - Support for notification balloons
 - Repository filtering
 
 ### Fixed
+
 - Fix nullable htmlUrl handling in DTO and related logic
 
 ## [0.0.1] - 2024-12-29
+
 Initial release of the GitHub Notifications IntelliJ Plugin.
 
 Done is better than perfect.
+
+[Unreleased]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.0.1...v0.2.0
+[0.0.1]: https://github.com/naoyukik/intellij-plugin-github-notifications/commits/v0.0.1